+++ SoftwareEngineeringProjekt2/src/dtu/library/app/LogElement.java
@@ -3,9 +3,9 @@ package dtu.library.app;
 public class LogElement {
 	
 	public Activity activity;
-	public int hours;
+	public float hours;
 	
-	public LogElement(Activity activity, int hours) {
+	public LogElement(Activity activity, float hours) {
 		this.activity = activity;
 		this.hours = hours;
 	}
